Split bridged room into a Channel and Webhook type
