+++ src/rooms/ChannelRoom.ts
@@ -1,0 +1,1093 @@
+/*
+Copyright 2019 The Matrix.org Foundation C.I.C.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import axios from "axios";
+import { Logging, Intent } from "matrix-appservice-bridge";
+import { SlackGhost } from "../SlackGhost";
+import { Main, METRIC_SENT_MESSAGES } from "../Main";
+import { default as substitutions, getFallbackForMissingEmoji } from "../substitutions";
+import * as emoji from "node-emoji";
+import { ISlackMessageEvent, ISlackEvent, ISlackFile } from "../BaseSlackHandler";
+import { WebAPIPlatformError, WebClient } from "@slack/web-api";
+import { ChatUpdateResponse,
+    ChatPostMessageResponse, ConversationsInfoResponse, FileInfoResponse } from "../SlackResponses";
+import { RoomEntry, EventEntry, TeamEntry } from "../datastore/Models";
+import { BridgedRoom, ISlackChatMessagePayload, MatrixEditEvent, MatrixMessageEvent, MatrixReactionEvent,
+    MatrixRedactionEvent, SlackChannelTypes, stripMatrixReplyFallback } from "../BridgedRoom";
+
+const log = Logging.get("ChannelRoom");
+interface ChannelRoomOpts {
+    matrixRoomId: string;
+    inboundId: string;
+    slackChannelId: string;
+    slackType: SlackChannelTypes;
+    isPrivate: boolean;
+    puppetOwner?: string;
+}
+
+const RECENT_MESSAGE_MAX = 10;
+const PUPPET_INCOMING_DELAY_MS = 1500;
+
+export class ChannelRoom extends BridgedRoom {
+    public get SlackChannelId(): string {
+        return this.slackChannelId;
+    }
+
+    public set SlackChannelId(value: string) {
+        this.dirty = this.dirty || (this.slackChannelId !== value);
+        this.SlackChannelId = value;
+    }
+
+    public get SlackTeamId(): string {
+        return this.team.id;
+    }
+
+    public get SlackClient(): WebClient {
+        return this.botClient;
+    }
+
+    public get IsPrivate(): boolean {
+        return this.isPrivate;
+    }
+
+    public get SlackType(): SlackChannelTypes {
+        return this.slackType;
+    }
+
+    private slackChannelId: string;
+    private slackType: SlackChannelTypes;
+    private isPrivate = false;
+    private puppetOwner?: string;
+
+    private intent?: Intent;
+    private recentSlackMessages: string[] = [];
+
+    private slackSendLock: Promise<void> = Promise.resolve();
+
+    constructor(main: Main, opts: ChannelRoomOpts, private team: TeamEntry, private botClient: WebClient) {
+        super(main, opts.matrixRoomId, opts.inboundId);
+        this.MatrixRoomActive = true;
+        this.slackChannelId = opts.slackChannelId;
+        this.slackType = opts.slackType || "channel";
+        this.isPrivate = opts.isPrivate;
+        this.puppetOwner = opts.puppetOwner;
+    }
+
+    public updateUsingChannelInfo(channelInfo: ConversationsInfoResponse): void {
+        const chan = channelInfo.channel;
+        this.isPrivate = chan.is_private || false;
+        if (chan.is_channel) {
+            this.slackType = "channel";
+        } else if (chan.is_mpim) {
+            this.slackType = "mpim";
+            // note: is_group is also set for mpims, so order is important
+        } else if (chan.is_group) {
+            this.slackType = "group";
+        } else if (chan.is_im) {
+            this.slackType = "im";
+        } else {
+            this.slackType = "unknown";
+        }
+        // Set this to true as we've changed the room.
+        this.dirty = true;
+    }
+
+    public getStatus(): string {
+        return "ready";
+    }
+
+    /**
+     * Returns data to write to the RoomStore
+     * As a side-effect will also clear the isDirty() flag
+     */
+    public toEntry(): RoomEntry {
+        const entry = {
+            id: `INTEG-${this.inboundId}`,
+            matrix_id: this.matrixRoomId,
+            remote: {
+                id: this.slackChannelId,
+                slack_team_id: this.team.id,
+                slack_type: this.slackType,
+                slack_private: this.isPrivate,
+                puppet_owner: this.puppetOwner,
+            },
+            remote_id: this.inboundId,
+        };
+        this.dirty = false;
+        return entry;
+    }
+
+    private async getClientForRequest(userId: string): Promise<{id: string, client: WebClient}|null> {
+        const puppet = await this.main.clientFactory.getClientForUserWithId(this.team.id, userId);
+        if (puppet) {
+            return puppet;
+        }
+        if (this.botClient) {
+            return {
+                id: "BOT",
+                client: this.botClient,
+            };
+        }
+        return null;
+    }
+
+    public async onMatrixReaction(message: MatrixReactionEvent): Promise<void> {
+        const relatesTo = message.content["m.relates_to"];
+        const eventStore = this.main.datastore;
+        const event = await eventStore.getEventByMatrixId(message.room_id, relatesTo.event_id);
+
+        // If we don't get an event then exit
+        if (event === null) {
+            log.debug("Could not find event to react to.");
+            return;
+        }
+
+        // Convert the unicode emoji into a slack emote name
+        let emojiKeyName: string;
+        const emojiItem = emoji.find(relatesTo.key);
+        if (emojiItem !== undefined) {
+            emojiKeyName = emojiItem.key;
+        } else {
+            emojiKeyName = relatesTo.key;
+            // Strip the colons
+            if (emojiKeyName.startsWith(":") && emojiKeyName.endsWith(":")) {
+                emojiKeyName = emojiKeyName.substring(1, emojiKeyName.length - 1);
+            }
+        }
+        const clientForRequest = await this.getClientForRequest(message.sender);
+        if (!clientForRequest) {
+            log.warn("No client to handle reaction");
+            return;
+        }
+        const { client, id } = clientForRequest;
+        // We must do this before sending to avoid racing
+        // Use the unicode key for uniqueness
+        this.addRecentSlackMessage(`reactadd:${relatesTo.key}:${id}:${event.slackTs}`);
+
+        // TODO: This only works once from matrix if we are sending the event as the
+        // bot user. Search for #fix_reactions_as_bot.
+        const res = await client.reactions.add({
+            as_user: false,
+            channel: this.slackChannelId,
+            name: emojiKeyName,
+            timestamp: event.slackTs,
+        });
+
+        if (!res.ok) {
+            log.error(`HTTP Error from Slack when adding the reaction :${emojiKeyName}: to ${event.slackTs}: `, res.error);
+            return;
+        }
+
+        log.info(`Reaction :${emojiKeyName}: added to ${event.slackTs}. Matrix room ID: ${message.room_id}. Matrix event ID: ${message.event_id}`);
+
+        await this.main.datastore.upsertReaction({
+            roomId: message.room_id,
+            eventId: message.event_id,
+            slackChannelId: this.slackChannelId,
+            slackMessageTs: event.slackTs,
+            // TODO We post reactions as the bot, not the user. Search for #fix_reactions_as_bot.
+            slackUserId: this.team.user_id,
+            reaction: emojiKeyName,
+        });
+    }
+
+    public async onMatrixRedaction(message: MatrixRedactionEvent): Promise<void> {
+        const clientForRequest = await this.getClientForRequest(message.sender);
+        if (!clientForRequest) {
+            log.warn("No client to handle redaction");
+            return;
+        }
+        const { client } = clientForRequest;
+
+        const event = await this.main.datastore.getEventByMatrixId(message.room_id, message.redacts);
+
+        // If we don't get an event then exit
+        if (event === null) {
+            const reactionEntry = await this.main.datastore.getReactionByMatrixId(message.room_id, message.redacts);
+
+            if (reactionEntry) {
+                await this.main.datastore.deleteReactionByMatrixId(message.room_id, message.redacts);
+                const reactionDescription = `"${reactionEntry.reaction}" on message ${reactionEntry.slackMessageTs} ` +
+                    `in channel ${reactionEntry.slackChannelId}`;
+                try {
+                    await client.reactions.remove({
+                        as_user: false,
+                        channel: reactionEntry.slackChannelId,
+                        timestamp: reactionEntry.slackMessageTs,
+                        name: reactionEntry.reaction,
+                    });
+                    log.info(`Redacted reaction ${reactionDescription}`);
+                } catch (error) {
+                    log.warn(`Failed redact reaction ${reactionDescription}. Matrix room/event: ${message.room_id}, ${message.redacts}`);
+                    log.warn(error);
+                    throw error;
+                }
+                return;
+            }
+
+            log.debug(`Could not find event '${message.redacts}' in room '${message.room_id}' to delete.`);
+            return;
+        }
+
+        // Delete event so it's not over-redacted on Matrix when we receive the "message_deleted" event from Slack.
+        // https://github.com/matrix-org/matrix-appservice-slack/issues/430
+        await this.main.datastore.deleteEventByMatrixId(message.room_id, message.redacts);
+
+        try {
+            // Note: bots can only delete their own messages, ergo it's possible that this may fail :(
+            await client.chat.delete({
+                as_user: false,
+                channel: this.slackChannelId,
+                ts: event.slackTs,
+            });
+        } catch (ex) {
+            log.warn(`Failed to delete ${event.slackTs}`, ex);
+        }
+    }
+
+    public async onMatrixEdit(message: MatrixEditEvent): Promise<boolean> {
+        const clientForRequest = await this.getClientForRequest(message.sender);
+        if (!clientForRequest) {
+            log.warn("No client to handle edit");
+            return false;
+        }
+        const { client } = clientForRequest;
+
+        const event = await this.main.datastore.getEventByMatrixId(
+            message.room_id,
+            message.content["m.relates_to"].event_id,
+        );
+
+        if (!event) {
+            log.debug("Skipping matrix edit because couldn't find event in datastore");
+            return false;
+        }
+        // re-write the message so the matrixToSlack converter works as expected.
+        let newMessage = JSON.parse(JSON.stringify(message));
+        newMessage.content = message.content["m.new_content"];
+        newMessage = stripMatrixReplyFallback(newMessage);
+
+        const body = await substitutions.matrixToSlack(newMessage, this.main, this.SlackTeamId);
+
+        if (!body || !body.text) {
+            log.warn(`Dropped edit ${message.event_id}, message content could not be identified`);
+            // Could not handle content, dropped
+            return false;
+        }
+
+        const res = (await client.chat.update({
+            ts: event.slackTs,
+            as_user: false,
+            channel: this.slackChannelId,
+            ...body,
+            // We include this for type safety as Typescript isn't aware that body.text is defined
+            // from the ...body statement.
+            text: body.text,
+        })) as ChatUpdateResponse;
+
+        this.main.incCounter(METRIC_SENT_MESSAGES, {side: "remote"});
+        if (!res.ok) {
+            log.error("HTTP Error: ", res.error);
+            throw Error("Failed to send");
+        }
+        // Add this event to the event store
+        await this.main.datastore.upsertEvent(
+            message.room_id,
+            message.event_id,
+            this.slackChannelId,
+            res.ts,
+        );
+        return true;
+    }
+
+    public async onMatrixMessage(message: MatrixMessageEvent): Promise<boolean> {
+        const puppetedClient = await this.main.clientFactory.getClientForUser(this.SlackTeamId, message.sender);
+        if (!this.botClient && !puppetedClient) { return false; }
+        const slackClient = puppetedClient || this.botClient;
+        const user = this.main.getOrCreateMatrixUser(message.sender);
+        message = stripMatrixReplyFallback(message);
+        const matrixToSlackResult = await substitutions.matrixToSlack(message, this.main, this.SlackTeamId);
+        if (!matrixToSlackResult) {
+            // Could not handle content, dropped.
+            log.warn(`Dropped ${message.event_id}, message content could not be identified`);
+            return false;
+        }
+        const body: ISlackChatMessagePayload = {
+            ...matrixToSlackResult,
+            as_user: false,
+            username: user.getDisplaynameForRoom(message.room_id) || matrixToSlackResult.username,
+        };
+        const text = body.text;
+        if (!body.attachments && !text) {
+            // The message type might not be understood. In any case, we can't send something without
+            // text.
+            log.warn(`Dropped ${message.event_id}, message had no attachments or text`);
+            return false;
+        }
+        const reply = await this.findParentReply(message);
+        let parentStoredEvent: EventEntry | null = null;
+        if (reply !== message.event_id) {
+            parentStoredEvent = await this.main.datastore.getEventByMatrixId(message.room_id, reply);
+            // We have a reply
+            if (parentStoredEvent) {
+                body.thread_ts = parentStoredEvent.slackTs;
+            }
+        }
+
+        const avatarUrl = user.getAvatarUrlForRoom(message.room_id);
+
+        if (avatarUrl && avatarUrl.indexOf("mxc://") === 0) {
+            body.icon_url = this.main.getUrlForMxc(avatarUrl);
+        }
+
+        user.bumpATime();
+        this.matrixATime = Date.now() / 1000;
+        if (puppetedClient) {
+            body.as_user = true;
+            delete body.username;
+        }
+        let res: ChatPostMessageResponse;
+        const chatPostMessageArgs = {
+            ...body,
+            // Ensure that text is defined, even for attachments.
+            text: text || "",
+            channel: this.slackChannelId,
+            unfurl_links: true,
+        };
+
+        try {
+            res = await slackClient.chat.postMessage(chatPostMessageArgs) as ChatPostMessageResponse;
+        } catch (ex) {
+            const platformError = ex as WebAPIPlatformError;
+            if (platformError.data?.error === "not_in_channel") {
+                await slackClient.conversations.join({
+                    channel: chatPostMessageArgs.channel,
+                });
+                res = await slackClient.chat.postMessage(chatPostMessageArgs) as ChatPostMessageResponse;
+            } else {
+                throw ex;
+            }
+        }
+
+        this.addRecentSlackMessage(res.ts);
+
+        this.main.incCounter(METRIC_SENT_MESSAGES, {side: "remote"});
+        // Log activity, but don't await the answer or throw errors
+        this.main.datastore.upsertActivityMetrics(user, this).catch((err) => {
+            log.error(`Error storing activity metrics`, err);
+        });
+
+        if (!res.ok) {
+            log.error("HTTP Error: ", res.error);
+            throw Error("Failed to send");
+        }
+
+        // Add this event to the event store
+        await this.main.datastore.upsertEvent(
+            message.room_id,
+            message.event_id,
+            this.slackChannelId,
+            res.ts,
+        );
+
+        // If this message is in a slack thread we need to append this message to the end of the thread list.
+        if (parentStoredEvent) {
+            if (parentStoredEvent._extras.slackThreadMessages === undefined) {
+                parentStoredEvent._extras.slackThreadMessages = [];
+            }
+            parentStoredEvent._extras.slackThreadMessages.push(res.ts);
+            await this.main.datastore.upsertEvent(parentStoredEvent);
+        }
+        return true;
+    }
+
+    public async onSlackUserLeft(slackId: string): Promise<void> {
+        const ghost = await this.main.ghostStore.get(slackId, undefined, this.SlackTeamId);
+        await ghost.intent.leave(this.matrixRoomId);
+    }
+
+    public async onSlackUserJoin(slackId: string, wasInvitedBy?: string): Promise<void> {
+        // There are different flows for this:
+        // 1 - Slack user invites slack user
+        // 2 - Slack user invites matrix user
+        // 3 - Matrix user invites slack user
+        // 4 - Matrix user invites matrix user
+        // 5 - Slack user has joined
+        // 6 - Matrix user has joined
+
+        if (this.team?.user_id === slackId) {
+            // We never reflect our own slack bot.
+            return;
+        }
+
+        const recipientPuppet = await this.main.clientFactory.getClientForSlackUser(this.SlackTeamId, slackId);
+        const recipientGhost = await this.main.ghostStore.get(slackId, undefined, this.SlackTeamId);
+
+        const senderPuppet = await this.main.clientFactory.getClientForSlackUser(this.SlackTeamId, slackId);
+        const senderGhost = await this.main.ghostStore.get(slackId, undefined, this.SlackTeamId);
+        const mxid = await this.main.datastore.getPuppetMatrixUserBySlackId(this.SlackTeamId, slackId);
+
+        if (!wasInvitedBy) {
+            if (!recipientPuppet) {
+                log.debug(`S-> ${slackId} joined ${this.SlackChannelId}`);
+                // 5
+                await recipientGhost.intent.join(this.matrixRoomId);
+            } else if (mxid) {
+                log.debug(`M-> ${slackId} joined ${this.SlackChannelId}`);
+                // 6
+                await this.main.botIntent.invite(this.matrixRoomId, mxid);
+            }
+            return;
+        }
+
+        if (!recipientPuppet && !senderPuppet) {
+            log.debug(`S->S ${slackId} was invited by ${wasInvitedBy}`);
+            // 1
+            await senderGhost.intent.invite(this.matrixRoomId, recipientGhost.userId);
+            await recipientGhost.intent.join(this.matrixRoomId);
+        } else if (recipientPuppet && mxid) {
+            // 2 & 4
+            log.debug(`S|M->M${mxid} was invited by ${wasInvitedBy}`);
+            await senderGhost.intent.invite(this.matrixRoomId, mxid);
+        } else if (senderPuppet && !recipientPuppet) {
+            log.debug(`M->S ${slackId} was invited by ${wasInvitedBy}`);
+            // 3
+            await recipientGhost.intent.join(this.matrixRoomId);
+            // No-op
+        }
+    }
+
+    public async onMatrixLeave(userId: string): Promise<void> {
+        log.info(`Leaving ${userId} from ${this.SlackChannelId}`);
+        const puppetedClient = await this.main.clientFactory.getClientForUser(this.SlackTeamId, userId);
+        if (!puppetedClient) {
+            log.debug("No client");
+            return;
+        }
+        await puppetedClient.conversations.leave({ channel: this.SlackChannelId });
+    }
+
+    public async onMatrixJoin(userId: string): Promise<void> {
+        log.info(`Joining ${userId} to ${this.SlackChannelId}`);
+        const puppetedClient = await this.main.clientFactory.getClientForUser(this.SlackTeamId, userId);
+        if (!puppetedClient) {
+            log.debug("No client");
+            return;
+        }
+        await puppetedClient.conversations.join({ channel: this.SlackChannelId });
+    }
+
+    public async onMatrixInvite(inviter: string, invitee: string): Promise<void> {
+        const puppetedClient = await this.main.clientFactory.getClientForUser(this.SlackTeamId, inviter);
+        if (!puppetedClient) {
+            log.debug("No client");
+            return;
+        }
+        const ghost = await this.main.ghostStore.get(invitee);
+        if (!ghost) {
+            log.debug("No ghost");
+            return;
+        }
+        await puppetedClient.conversations.invite({channel: this.slackChannelId, users: ghost.slackId });
+    }
+
+    public async onSlackMessage(message: ISlackMessageEvent): Promise<void> {
+        if (message.user) {
+            // This just checks if the user *could* be puppeted. If they are, delay handling their incoming messages.
+            const hasPuppet = null !== await this.main.datastore.getPuppetTokenBySlackId(this.SlackTeamId, message.user);
+            if (hasPuppet) {
+                await new Promise((r) => setTimeout(r, PUPPET_INCOMING_DELAY_MS));
+            }
+        }
+        if (this.recentSlackMessages.includes(message.ts)) {
+            // We sent this, ignore.
+            return;
+        }
+        try {
+            const ghost = await this.main.ghostStore.getForSlackMessage(message, this.SlackTeamId);
+            await ghost.update(message, this);
+            await ghost.cancelTyping(this.MatrixRoomId); // If they were typing, stop them from doing that.
+            this.slackSendLock = this.slackSendLock.finally(async () => {
+                // Check again
+                if (!this.recentSlackMessages.includes(message.ts)) {
+                    return this.handleSlackMessage(message, ghost);
+                }
+                // We sent this, ignore
+            });
+            await this.slackSendLock;
+        } catch (err) {
+            log.error("Failed to process event");
+            log.error(err);
+        }
+    }
+
+    public async onSlackReactionAdded(
+        message: {
+            event_ts: string,
+            item: {
+                channel: string,
+                ts: string,
+            },
+            reaction: string,
+            user_id: string,
+        },
+        teamId: string,
+    ): Promise<void> {
+        if (message.user_id === this.team.user_id) {
+            return;
+        }
+
+        let reactionKey = emoji.emojify(`:${message.reaction}:`, getFallbackForMissingEmoji);
+        // Element uses the default thumbsup and thumbsdown reactions with an appended variant character.
+        if (reactionKey === '👍' || reactionKey === '👎') {
+            reactionKey += '\ufe0f'.normalize(); // VARIATION SELECTOR-16
+        }
+
+        if (this.recentSlackMessages.includes(`reactadd:${reactionKey}:${message.user_id}:${message.item.ts}`)) {
+            // We sent this, ignore.
+            return;
+        }
+        const ghost = await this.main.ghostStore.getForSlackMessage(message, teamId);
+        await ghost.update(message, this);
+
+        const event = await this.main.datastore.getEventBySlackId(message.item.channel, message.item.ts);
+
+        if (event === null) {
+            return;
+        }
+        let response: { event_id: string };
+        const reactionDesc = `${reactionKey} for ${event.eventId} as ${ghost.userId}. Matrix room/event: ${this.MatrixRoomId}, ${event.eventId}`;
+        try {
+            response = await ghost.sendReaction(
+                this.MatrixRoomId,
+                event.eventId,
+                reactionKey,
+                message.item.channel,
+                message.event_ts
+            );
+            log.info(`Sending reaction ${reactionDesc}`);
+        } catch (error) {
+            log.warn(`Failed to send reaction ${reactionDesc}`);
+            throw error;
+        }
+        await this.main.datastore.upsertReaction({
+            roomId: this.MatrixRoomId,
+            eventId: response.event_id,
+            slackChannelId: message.item.channel,
+            slackMessageTs: message.item.ts,
+            slackUserId: message.user_id,
+            reaction: message.reaction,
+        });
+    }
+
+    public async onSlackReactionRemoved(
+        msg: {
+            item: {
+                channel: string,
+                ts: string,
+            },
+            reaction: string,
+            user_id: string,
+        },
+    ): Promise<void> {
+        if (msg.user_id === this.team.user_id) {
+            return;
+        }
+        const originalEvent = await this.main.datastore.getReactionBySlackId(msg.item.channel, msg.item.ts, msg.user_id, msg.reaction );
+        if (!originalEvent) {
+            throw Error('unknown_reaction');
+        }
+        const botClient = this.main.botIntent.getClient();
+        botClient.redactEvent(originalEvent.roomId, originalEvent.eventId);
+        await this.main.datastore.deleteReactionBySlackId(msg.item.channel, msg.item.ts, msg.user_id, msg.reaction);
+    }
+
+    public async onSlackTyping(event: ISlackEvent, teamId: string): Promise<void> {
+        const ghost = await this.main.ghostStore.getForSlackMessage(event, teamId);
+        await ghost.sendTyping(this.MatrixRoomId);
+    }
+
+    public setBotClient(slackClient: WebClient): void {
+        this.botClient = slackClient;
+    }
+
+    private async handleSlackMessageFile(file: ISlackFile, slackEventId: string, ghost: SlackGhost) {
+        const maxUploadSize = this.main.config.homeserver.max_upload_size;
+        const filePrivateUrl = file.url_private;
+        if (!filePrivateUrl) {
+            log.info(`Slack file ${file.id} lacks a url_private, not handling file.`);
+            return;
+        }
+        const channelId = this.slackChannelId;
+        if (!channelId) {
+            // The ID is required.
+            return;
+        }
+
+        let sendAsLink = false;
+        const authToken = this.SlackClient?.token;
+
+        if (!authToken) {
+            log.error("We have no client (or token) that can handle this file, sending as link");
+            sendAsLink = true;
+        } else if (maxUploadSize && file.size > maxUploadSize) {
+            log.warn(`File size too large (${file.size / 1024}KiB > ${maxUploadSize / 1024} KB)`);
+            sendAsLink = true;
+        }
+
+        if (sendAsLink) {
+            const link = file.public_url_shared ? file.permalink_public : file.url_private;
+            const messageContent = {
+                body: `${link} (${file.name})`,
+                format: "org.matrix.custom.html",
+                formatted_body: `<a href="${link}">${file.name}</a>`,
+                msgtype: "m.text",
+            };
+            await ghost.sendMessage(this.matrixRoomId, messageContent, channelId, slackEventId);
+            return;
+        }
+
+        if (!authToken) {
+            // To ensure that Typescript is happy
+            return;
+        }
+
+        if (file.mode === "snippet") {
+            let htmlString: string;
+            try {
+                const fileReq = await axios.get<string>(filePrivateUrl, {
+                    headers: {
+                        // Token is checked above.
+                        Authorization: `Bearer ${authToken}`,
+                    }
+                });
+                if (fileReq.status !== 200) {
+                    // We don't want to accidentally publish a error page.
+                    throw Error('Non-200 status returned for snippet');
+                }
+                htmlString = fileReq.data;
+            } catch (ex) {
+                log.error("Failed to download snippet", ex);
+                return;
+            }
+            let htmlCode = "";
+            // Because escaping 6 backticks is not good for readability.
+            const code = "```" + `\n${htmlString}\n` + "```";
+            if (file.filetype) {
+                htmlCode = `<pre><code class="language-${file.filetype}'">`;
+            } else {
+                htmlCode = "<pre><code>";
+            }
+            htmlCode += substitutions.htmlEscape(htmlString);
+            htmlCode += "</code></pre>";
+
+            const messageContent = {
+                body: code,
+                format: "org.matrix.custom.html",
+                formatted_body: htmlCode,
+                msgtype: "m.text",
+            };
+            await ghost.sendMessage(this.matrixRoomId, messageContent, channelId, slackEventId);
+            return;
+        }
+
+        // Sometimes Slack sends us a message too soon, and the file is missing it's mimetype.
+        if (!file.mimetype) {
+            log.info(`Slack file ${file.id} is missing mimetype, fetching fresh info`);
+            file = ((await this.SlackClient.files.info({
+                file: file.id,
+            })) as FileInfoResponse).file;
+            // If it's *still* missing a mimetype, we'll treat it as a file later.
+        }
+
+        // We also need to upload the thumbnail
+        let thumbnailPromise: Promise<string> = Promise.resolve("");
+        // Slack ain't a believer in consistency.
+        const thumbUri = file.thumb_video || file.thumb_360;
+        if (thumbUri && file.filetype && this.SlackClient.token) {
+            thumbnailPromise = ghost.uploadContentFromURI(
+                {
+                    mimetype: file.mimetype,
+                    title: `${file.name}_thumb.${file.filetype}`,
+                },
+                thumbUri,
+                authToken,
+            );
+        }
+
+        const fileContentUri = await ghost.uploadContentFromURI(
+            // authToken is verified above.
+            file, filePrivateUrl, authToken);
+        const thumbnailContentUri = await thumbnailPromise;
+        await ghost.sendMessage(
+            this.matrixRoomId,
+            slackFileToMatrixMessage(file, fileContentUri, thumbnailContentUri),
+            channelId,
+            slackEventId,
+        );
+    }
+
+    private async handleSlackMessage(message: ISlackMessageEvent, ghost: SlackGhost) {
+        const eventTS = message.event_ts || message.ts;
+        const channelId = this.slackChannelId;
+
+        // Dedupe across RTM/Event streams
+        this.addRecentSlackMessage(message.ts);
+
+        ghost.bumpATime();
+        this.slackATime = Date.now() / 1000;
+
+        const subtype = message.subtype;
+
+        // Log activity, but don't await the answer or throw errors
+        this.main.datastore.upsertActivityMetrics(ghost, this).catch((err) => {
+            log.error(`Error storing activity metrics`, err);
+        });
+
+        // Transform the text if it is present.
+        if (message.text) {
+            message.text = substitutions.slackToMatrix(message.text,
+                subtype === "file_comment" ? message.file : undefined);
+        }
+
+        if (message.thread_ts !== undefined && message.text) {
+            let replyMEvent = await this.getReplyEvent(this.MatrixRoomId, message, this.SlackChannelId);
+            if (replyMEvent) {
+                replyMEvent = stripMatrixReplyFallback(replyMEvent);
+                return await ghost.sendWithReply(
+                    this.MatrixRoomId, message.text, this.SlackChannelId, eventTS, replyMEvent,
+                );
+            } else {
+                log.warn("Could not find matrix event for parent reply", message.thread_ts);
+            }
+        }
+
+        // If we are only handling text, send the text. File messages are handled in a seperate block.
+        if (["bot_message", "file_comment", undefined].includes(subtype) && message.files === undefined) {
+            return ghost.sendText(this.matrixRoomId, message.text!, channelId, eventTS);
+        } else if (subtype === "me_message") {
+            return ghost.sendMessage(this.matrixRoomId, {
+                body: message.text!,
+                msgtype: "m.emote",
+            }, channelId, eventTS);
+        } else if (subtype === "message_changed") {
+            const previousMessage = ghost.prepareBody(substitutions.slackToMatrix(message.previous_message!.text!));
+            // We use message.text here rather than the proper message.message.text
+            // as we have added message.text ourselves and then transformed it.
+            const newMessageRich = substitutions.slackToMatrix(message.text!);
+            const newMessage = ghost.prepareBody(newMessageRich);
+
+            // The substitutions might make the messages the same
+            if (previousMessage === newMessage) {
+                log.debug("Ignoring edit message because messages are the same post-substitutions.");
+                return;
+            }
+
+            const edits = substitutions.makeDiff(previousMessage, newMessage);
+
+            const outtext = `(edited) ${edits.before} ${edits.prev} ${edits.after} => ` +
+                `${edits.before} ${edits.curr}  ${edits.after}`;
+
+            const prev   = substitutions.htmlEscape(edits.prev);
+            const curr   = substitutions.htmlEscape(edits.curr);
+            const before = substitutions.htmlEscape(edits.before);
+            const after  = substitutions.htmlEscape(edits.after);
+
+            let formatted = `<i>(edited)</i> ${before} <font color="red"> ${prev} </font> ${after} =&gt; ${before}` +
+            `<font color="green"> ${curr} </font> ${after}`;
+            const prevEvent = await this.main.datastore.getEventBySlackId(channelId, message.previous_message!.ts);
+
+            // If this edit is in a thread we need to inject the reply fallback, or
+            // non-reply supporting clients will no longer show it as a reply.
+            let body = ghost.prepareBody(outtext);
+
+            let newBody = ghost.prepareBody(newMessageRich);
+            let newFormattedBody = ghost.prepareFormattedBody(newMessageRich);
+            if (message.message && message.message.thread_ts !== undefined) {
+                let replyEvent = await this.getReplyEvent(
+                    this.MatrixRoomId, message.message as unknown as ISlackMessageEvent, this.slackChannelId,
+                );
+                replyEvent = stripMatrixReplyFallback(replyEvent);
+                if (replyEvent) {
+                    const bodyFallback = ghost.getFallbackText(replyEvent);
+                    const formattedFallback = ghost.getFallbackHtml(this.MatrixRoomId, replyEvent);
+                    body = `${bodyFallback}\n\n${body}`;
+                    formatted = formattedFallback + formatted;
+                    newBody = bodyFallback + newBody;
+                    newFormattedBody = formattedFallback + newFormattedBody;
+                }
+            }
+            let replyContent: Record<string, unknown>|undefined;
+            // Only include edit metadata in the message if we have the previous eventId,
+            // otherwise just send the fallback reply text.
+            if (prevEvent) {
+                replyContent = {
+                    "m.new_content": {
+                        body: newBody,
+                        format: "org.matrix.custom.html",
+                        formatted_body: newFormattedBody,
+                        msgtype: "m.text",
+                    },
+                    "m.relates_to": {
+                        event_id: prevEvent.eventId,
+                        rel_type: "m.replace",
+                    },
+                };
+            } else {
+                log.warn("Got edit but no previous matrix events were found");
+            }
+            const matrixContent = {
+                body,
+                format: "org.matrix.custom.html",
+                formatted_body: formatted,
+                msgtype: "m.text",
+                ...replyContent,
+            };
+            return ghost.sendMessage(this.MatrixRoomId, matrixContent, channelId, eventTS);
+        } else if (message.files) { // A message without a subtype can contain files.
+            for (const file of message.files) {
+                try {
+                    await this.handleSlackMessageFile(file, eventTS, ghost);
+                } catch (ex) {
+                    log.warn(`Couldn't handle Slack file, ignoring:`, ex);
+                }
+            }
+            // TODO: Currently Matrix lacks a way to upload a "captioned image",
+            //   so we just send a separate `m.image` and `m.text` message
+            // See https://github.com/matrix-org/matrix-doc/issues/906
+            if (message.text) {
+                return ghost.sendText(this.matrixRoomId, message.text, channelId, eventTS);
+            }
+        } else {
+            log.warn(`Ignoring message with subtype: ${subtype}`);
+        }
+    }
+
+    private async getReplyEvent(roomID: string, message: ISlackMessageEvent, slackRoomID: string) {
+        if (!message.thread_ts) {
+            return;
+        }
+        // Get parent event
+        const dataStore = this.main.datastore;
+        const parentEvent = await dataStore.getEventBySlackId(slackRoomID, message.thread_ts);
+        if (parentEvent === null) {
+            log.warn(`Could not find parent matrix event for ${message.thread_ts}`);
+            return null;
+        }
+        let replyToTS = "";
+        // Add this event to the list of events in this thread
+        if (parentEvent._extras.slackThreadMessages === undefined) {
+            parentEvent._extras.slackThreadMessages = [];
+        }
+        replyToTS = parentEvent._extras.slackThreadMessages.slice(-1)[0] || message.thread_ts;
+        parentEvent._extras.slackThreadMessages.push(message.ts);
+        await dataStore.upsertEvent(parentEvent);
+
+        // Get event to reply to
+        const replyToEvent = await dataStore.getEventBySlackId(slackRoomID, replyToTS);
+        if (replyToEvent === null) {
+            log.warn(`Could not find parent matrix event for the latest event in the chain ${replyToTS}`);
+            return null;
+        }
+        const intent = await this.getIntentForRoom(roomID);
+        return await intent.getClient().fetchRoomEvent(roomID, replyToEvent.eventId);
+    }
+
+    /*
+        Given an event which is in reply to something else return the event ID of the
+        top most event in the reply chain, i.e. the one without a relates to.
+    */
+    private async findParentReply(message: any, depth = 0): Promise<string> {
+        const MAX_DEPTH = 10;
+        // Extract the referenced event
+        if (!message.content) { return message.event_id; }
+        if (!message.content["m.relates_to"]) { return message.event_id; }
+        if (!message.content["m.relates_to"]["m.in_reply_to"]) { return message.event_id; }
+        const parentEventId = message.content["m.relates_to"]["m.in_reply_to"].event_id;
+        if (!parentEventId) { return message.event_id; }
+        if (depth > MAX_DEPTH) {
+            return parentEventId; // We have hit our depth limit, use this one.
+        }
+
+        const intent = await this.getIntentForRoom(message.room_id);
+        const nextEvent = await intent.getClient().fetchRoomEvent(message.room_id, parentEventId);
+
+        return this.findParentReply(nextEvent, depth++);
+    }
+
+    private async getIntentForRoom(roomID: string) {
+        if (this.intent) {
+            return this.intent;
+        }
+        // Ensure we get the right user.
+        if (!this.IsPrivate) {
+            this.intent = this.main.botIntent; // Non-private channels should have the bot inside.
+        }
+        const firstGhost = (await this.main.listGhostUsers(roomID))[0];
+        this.intent =  this.main.getIntent(firstGhost);
+        return this.intent;
+    }
+
+    private addRecentSlackMessage(ts: string) {
+        log.debug("Recent message key add:", ts);
+        this.recentSlackMessages.push(ts);
+        if (this.recentSlackMessages.length > RECENT_MESSAGE_MAX) {
+            this.recentSlackMessages.shift();
+        }
+    }
+}
+
+/**
+ * Converts a slack image attachment to a matrix image event.
+ *
+ * @param {Object} file The slack image attachment file object.
+ * @param {?integer} file.size size of the file in bytes.
+ * @param {string} file.title alt-text for the file.
+ * @param {string} file.mimetype mime-type of the file.
+ * @param {?integer} file.original_w width of the file if an image, in pixels.
+ * @param {?integer} file.original_h height of the file if an image, in pixels.
+ * @param {?string} file.thumb_360 URL of a 360 pixel wide thumbnail of the
+ * file, if an image.
+ * @param {?integer} file.thumb_360_w width of the thumbnail of the 360 pixel
+ * wide thumbnail of the file, if an image.
+ * @param {?integer} file.thumb_360_h height of the thumbnail of the 36 pixel
+ * wide thumbnail of the file, if an image.
+ * @param {string} url The matrix file mxc.
+ * @param {?string} thumbnail_url The matrix thumbnail mxc.
+ * @return {Object} Matrix event content, as per https://matrix.org/docs/spec/#m-image
+ */
+const slackImageToMatrixImage = (file, url: string, thumbnailUrl?: string) => {
+    const message = {
+        body: file.title,
+        info: {
+            mimetype: file.mimetype,
+            size: file.size,
+        },
+        msgtype: "m.image",
+        url,
+        // TODO: Define some matrix types
+    } as any;
+
+    if (file.original_w) {
+        message.info.w = file.original_w;
+    }
+
+    if (file.original_h) {
+        message.info.h = file.original_h;
+    }
+
+    if (thumbnailUrl) {
+        message.thumbnail_url = thumbnailUrl;
+        message.thumbnail_info = {};
+        if (file.thumb_360_w) {
+            message.thumbnail_info.w = file.thumb_360_w;
+        }
+        if (file.thumb_360_h) {
+            message.thumbnail_info.h = file.thumb_360_h;
+        }
+    }
+    return message;
+};
+
+/**
+ * Converts a slack video attachment to a matrix video event.
+ *
+ * @param file The slack video attachment file object.
+ * @param file.size size of the file in bytes.
+ * @param file.title alt-text for the file.
+ * @param file.mimetype mime-type of the file.
+ * @param file.original_w width of the file if an image, in pixels.
+ * @param file.original_h height of the file if an image, in pixels.
+ * @param url The matrix file mxc.
+ * @param thumbnail_url The matrix thumbnail mxc.
+ * @return Matrix event content, as per https://matrix.org/docs/spec/client_server/r0.4.0.html#m-video
+ */
+const slackImageToMatrixVideo = (file, url: string, thumbnailUrl?: string) => {
+    const message = {
+        body: file.title,
+        info: {
+            mimetype: file.mimetype,
+            size: file.size,
+        },
+        msgtype: "m.video",
+        url,
+        // TODO: Define some matrix types
+    } as any;
+
+    if (file.original_w) {
+        message.info.w = file.original_w;
+    }
+
+    if (file.original_h) {
+        message.info.h = file.original_h;
+    }
+
+    if (thumbnailUrl) {
+        message.thumbnail_url = thumbnailUrl;
+        // Slack don't tell us the thumbnail size for videos. Boo
+    }
+
+    return message;
+};
+
+/**
+ * Converts a slack audio attachment to a matrix audio event.
+ *
+ * @param {Object} file The slack audio attachment file object.
+ * @param {?integer} file.size size of the file in bytes.
+ * @param {string} file.title alt-text for the file.
+ * @param {string} file.mimetype mime-type of the file.
+ * @param {string} url The matrix file mxc.
+ * @return {Object} Matrix event content, as per https://matrix.org/docs/spec/client_server/r0.4.0.html#m-audio
+ */
+const slackImageToMatrixAudio = (file, url: string) => ({
+    body: file.title,
+    info: {
+        mimetype: file.mimetype,
+        size: file.size,
+    },
+    msgtype: "m.audio",
+    url,
+});
+/**
+ * Converts a slack file upload to a matrix file upload event.
+ *
+ * @param file The slack file object.
+ * @param url The matrix file mxc.
+ * @param thumbnail_url The matrix thumbnail mxc.
+ * @return Matrix event content, as per https://matrix.org/docs/spec/#m-file
+ */
+const slackFileToMatrixMessage = (file, url: string, thumbnailUrl?: string) => {
+    if (file.mimetype) {
+        if (file.mimetype.startsWith("image/")) {
+            return slackImageToMatrixImage(file, url, thumbnailUrl);
+        } else if (file.mimetype.startsWith("video/")) {
+            return slackImageToMatrixVideo(file, url, thumbnailUrl);
+        } else if (file.mimetype.startsWith("audio/")) {
+            return slackImageToMatrixAudio(file, url);
+        }
+    }
+
+    return  {
+        body: file.title,
+        info: {
+            mimetype: file.mimetype,
+            size: file.size,
+        },
+        msgtype: "m.file",
+        url,
+    };
+};
